Django/drfintro/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.0.0
Django==4.1
django-cors-headers==4.2.0
django-environ==0.10.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Faker==18.10.1
gunicorn==20.1.0
idna==3.4
kombu==5.3.1
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==4.6.0
requestes==0.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
